add ability to keep downloaded files archives (not delete after downloading and unzipping)

+++ ideaSupport/src/main/scala/org/jetbrains/sbtidea/download/idea/IdeaDistInstaller.scala
@@ -2,6 +2,7 @@ package org.jetbrains.sbtidea.download.idea
 
 
 import org.jetbrains.sbtidea.download.api.*
+import org.jetbrains.sbtidea.download.idea.IdeaDistInstaller.KeepDownloadedFilesVmOption
 import org.jetbrains.sbtidea.download.{BuildInfo, FileDownloader, IdeaUpdater, NioUtils}
 import org.jetbrains.sbtidea.{pathToPathExt, PluginLogger as log}
 
@@ -63,7 +64,10 @@ class IdeaDistInstaller(buildInfo: BuildInfo) extends Installer[IdeaDist] {
 
     fixAccessRights(ctx.baseDirectory)
 
-    NioUtils.delete(artifact)
+    val keepDownloadedFiles = System.getProperty(KeepDownloadedFilesVmOption) != null
+    if (!keepDownloadedFiles) {
+      NioUtils.delete(artifact)
+    }
     log.info(s"Installed ${buildInfo.edition.name}($buildInfo) to ${ctx.baseDirectory}")
     ctx.baseDirectory
   }
@@ -85,4 +89,8 @@ class IdeaDistInstaller(buildInfo: BuildInfo) extends Installer[IdeaDist] {
       }
     }
 
+}
+
+object IdeaDistInstaller {
+  val KeepDownloadedFilesVmOption = "sbt.idea.plugin.keep.downloaded.files"
 }
